+++ src/redux/containers/simulation.js
@@ -1,6 +1,6 @@
 import { connect } from "react-redux";
 import Simulation from "../../components/simulation/simulation";
-import { fetchSimulation } from "../actions";
+import fetchSimulation from "../actions";
 
 const mapStateToProps = state => state.simulation;
