+++ src/redux/reducers/index.js
@@ -1,6 +1,8 @@
 import { combineReducers } from "redux";
 import simulation from "./simulation";
+import strategyFrom from "./strategy-form";
 
 export default combineReducers({
-  simulation
+  simulation,
+  strategyFrom
 });
